services/notification-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
slowapi==0.1.9
httpx==0.27.0
celery==5.4.0
redis==5.0.4
twilio==9.1.0
sendgrid==6.11.0
jinja2==3.1.4
